requirements.txt
# Manipulação e Análise de Dados
pandas==3.0.2
duckdb==1.5.2
pyarrow==24.0.0  # Essencial para a integração de alta performance entre DuckDB e Pandas

# Suporte a arquivos Excel
openpyxl==3.1.5

# Integração com Google Drive e Google Sheets
google-api-python-client==2.193.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1

# Gestão de Variáveis de Ambiente (Segurança de Credenciais)
python-dotenv==1.2.2